service/requirements.txt
gunicorn==21.2.0
uvicorn==0.29.0
fastapi==0.110.1
python-dotenv==1.0.1
orjson==3.10.5
